+++ Priyanshu_Ranjan/.gitignore
@@ -1,4 +1,11 @@
+.env
 venv
-static
-uploads
-results.json
+ByteTrack
+images
+labels
+runs
+test_images
+classes.txt
+evaluation_metrics.txt
+testing.py
+upload_results.py
