--- Priyanshu_Ranjan/Web App/requirements.txt
@@ -1,15 +1,0 @@
-streamlit==1.27.2
-numpy==1.26.4
-pandas==2.2.2
-matplotlib==3.8.4
-
-# OpenCV (headless for servers)
-opencv-python-headless==4.9.0.80
-
-# Deep learning stack
-torch==2.2.2
-torchvision==0.17.2
-
-# YOLO + helpers
-ultralytics==8.1.47
-supervision==0.19.0
