--- Priyanshu_Ranjan/Web App/requirements_streamlit.txt
@@ -1,3 +1,0 @@
-streamlit==1.27.2
-opencv-python-headless==4.8.1.78
-numpy==1.26.4
