+++ Priyanshu_Ranjan/requirements.txt
@@ -1,0 +1,9 @@
+streamlit
+numpy
+pandas
+matplotlib
+opencv-python-headless
+torch
+torchvision
+ultralytics
+supervision
